Allow control of diagnostic sample size with num_samples flag
Previously, the number of synthetic samples generated during the
pauses in the training process (for computing stats, making plots etc)
were always 2**13. At every pause in the training process, this fake
sample was saved in the samples_fake.npz archive, overwriting it. At
the end of the training, a final set of fake samples was
generated, *this time of size FLAGS.num_samples*, and overwritten
again onto samples_fake.npz. So if you were to train the model you
would see the samples_fake.npz archive be of size 2**13 during
training, and then become of size FLAGS.num_samples just at the very
end, when training was complete.

This change should ensure that the size of this fake sample, used only
for diagnostic purposes, is always the same throughout the training
process.